requirements.txt
ipykernel
numpy
pandas
matplotlib
seaborn
flask
memory_profiler
streamlit
scikit-learn
langchain
python-dotenv
langchain-openai
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
ollama
sentence-transformers
langchain-huggingface
torch
tqdm
faiss-cpu
llama-cpp-python
langchain-ollama